Create Dummy target, make it spin around, assign layers
create a dummy target for debugging purposes, and make it orbit the player
at a somewhat constant speed, with the option to get closer as time goes
on.
also assign collision layers and masks to all objects, as well as change ship to an
area2d so that it can check for any fauna that get too close.